requirements.txt
asgiref==3.11.1
attrs==26.1.0
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cloudinary==1.44.1
dj-database-url==3.1.2
Django==5.2.12
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==25.1.0
h11==0.16.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
whitenoise==6.12.0
